backend/requirements.txt
fastapi==0.109.0
python-multipart==0.0.6
pymupdf==1.23.8
faiss-cpu==1.7.4
numpy==1.24.3
requests==2.31.0
pydantic==2.5.3
uvicorn==0.27.0